Rename GennyTest to Actor
