+++ src/utils/Actor.js
@@ -1,54 +1,81 @@
 import { SECONDS } from 'constants';
 import { setDefaultOptions } from 'expect-puppeteer';
 
-// Set maximum wait time for any element to 5 minutes
-setDefaultOptions({ timeout: 300 * SECONDS });
+// Set maximum wait time for any element to 3 minutes
+setDefaultOptions({ timeout: 240 * SECONDS });
+
+
+// jest and their methods are automatically defined
+jest.setTimeout( 100 * SECONDS );
 
 class Actor {
   constructor( page ) {
     this.page = page;
   }
 
-  async __fillTextBox( elementSelector, textValue ) {
-    await expect( this.page ).toFill( elementSelector, textValue );
+  async typeInput( inputType, askId, text, options = {}) {
+    const {
+      waitTime = 5,
+    } = options;
+
+    await this.page.waitFor( waitTime * SECONDS );
+
+    // Type into an input field on the page
+
+    await expect( this.page ).toFill( `[data-testid="input-${inputType} ${askId}"]`, text );
   }
 
-  async __clickButton( elementSelector ) {
-    await expect( this.page ).toClick( elementSelector );
+  async typeInputText( askId, text, options = {}) {
+    // Type into an text input on the page
+    await this.typeInput( 'text', askId, text, options );
   }
 
-  // Click an event button with a button code in it
-  async clickEventButton( buttonCode ) {
-    await this.__clickButton( `[data-testid="event-button ${buttonCode}"]` );
+  async typeInputEmail( askId, text, options = {}) {
+    // Type into an email input on the page
+    await this.typeInput( 'email', askId, text, options );
   }
 
-  // Click form button (either 'next' or 'submit')
-  async clickFormButton() {
-    await expect( this.page ).toClick( 'div[data-testid="button"]' );
+  async typeInputAutocomplete( askId, text ) {
+    // Type into the autocmplete input
+    await this.typeInput( askId, text );
+
+    // Click on the first autocomplete result
+    await expect( this.page ).toClick( `[data-testid="input-autocomplete-item ${askId}"]` );
   }
 
-  // Enter address by partially filling the textbox, and then clicking on the first autofill option.
-  async fillAddress( questionCode, answerValue ) {
-    await expect( this.page ).toFill(
-      `[data-testid="input-text ${questionCode}"]`,
-      answerValue
-    );
-    await expect( this.page ).toClick(
-      `[data-testid="input-autocomplete-item ${questionCode}"]`
-    );
+  async selectInput( askId, baseEntityCode ) {
+    // Find the dropdown input on the page and select the baseEntityCode from the items
+    await this.page.select( `select[data-testid="input-dropdown ${askId}"]`, baseEntityCode );
   }
 
-  // Fill in a generic text box
-  async fillTextBox( questionCode, answerValue ) {
-    this.__fillTextBox( `[data-testid="input-text ${questionCode}"]`, answerValue );
+  async click( testId, options = {}) {
+    const {
+      waitTime = 3,
+      clickIndex = 0,
+    } = options;
+
+    await this.page.waitFor( waitTime * SECONDS );
+
+    // Find the button the page
+    const button = this.page.$$( `[data-testid="${testId}"]` );
+
+    // Click the element of index `clickIndex`
+    await button[clickIndex].click();
   }
 
-  async fillDropdown( questionCode, optionValue ) {
-    await this.page.waitForSelector( `option[value="${optionValue}"]` );
-    await this.page.select(
-      `select[data-testid="input-dropdown ${questionCode}"]`,
-      optionValue
-    );
+  async clickButton( testId = '', options = {}) {
+    // Make a normal click but prefix it with `button`
+    await this.click( `button ${testId}`, options );
+  }
+
+  async clickDropdown( testId = '', options = {}) {
+    // Make a normal click but prefix it with `dropdown`
+    await this.click( `dropdown ${testId}`, options );
+  }
+
+  async clickDropdownItem( testId = '', options = {}) {
+    // Make a normal click but prefix it with `dropdown-item`
+    await this.click( `dropdown-item ${testId}`, options );
   }
 }
