+++ src/utils/index.js
@@ -1,11 +1,9 @@
 import { Screenshot } from './Screenshot';
-import { Actor } from './Actor';
-import GennyTest from './GennyTest';
+import Actor from './Actor';
 import One from './One';
 
 export {
   Screenshot,
   Actor,
-  GennyTest,
   One,
 };
